Rename benchmarkFunction to benchmark
